% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_pic.R
\name{draw_key_pic}
\alias{draw_key_pic}
\title{Draw key function for geom_bar_pic}
\usage{
draw_key_pic(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
To be able to generate a legend with the raster image, ggpic needs a custom
\code{draw_key} function.
}
\details{
When pic and fill aesthetics does not map to the same variable, fill guide
is redirected to \code{draw_key_rect} and pic guide to a raster grob.
}

