% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_pic.R
\name{geom_bar_pic}
\alias{geom_bar_pic}
\title{Pics as bars}
\usage{
geom_bar_pic(mapping = NULL, data = NULL, stat = "count",
  position = "dodge", width = NULL, na.rm = FALSE, show.legend = NA,
  asis = FALSE, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{Override the default connection between \code{geom_bar_pic}
and \code{stat_count}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{width}{Pic width. By default, set to 90\% of the resolution of the data.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{asis}{Logical indicating if the pic must be used as is, with no filling.
If FALSE, the default, image not transparent pixels will be filled to the
color provided in \code{fill} or black (default). If TRUE, image will
maintain the original colors.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
As in \code{\link[ggplot2]{geom_bar}}, depending on what is maped to bar_pic
height two types of bar_pic charts are produced. Default is \code{stat = "count"},
which makes the height of the bar_pic proportional to the numbers of cases in
each group. \code{stat = "identity"} will map height to values in the data,
so you need to map an \code{y} aesthetic also.
}
\details{
As opposed to \code{\link[ggplot2]{geom_bar}}, multiple x's occurring in the
same place will be "dodged" with \code{position = "dodge"}. If you want your
pics stacked, change to \code{position = "stack"}.
}
\section{Aesthetics}{

\code{geom_bar_pic} understand the following aesthetics (required aesthetics
are in bold):
\itemize{
    \item{\bold{x}}
    \item{pic}
    \item{alpha}
    \item{fill}
    \item{colour}
    \item{size}
    \item{linetype}
}
}

\section{Computed variables}{

\describe{
  \item{count}{number of points in bin}
  \item{prop}{groupwise proportion}
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}} in the \code{ggplot2} package
}

