% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_pic.R
\name{geom_point_pic}
\alias{geom_point_pic}
\title{Pics as points}
\usage{
geom_point_pic(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", width = NULL, na.rm = FALSE, show.legend = NA,
  asis = FALSE, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{asis}{Logical indicating if the pic must be used as is, with no filling.
If FALSE, the default, image not transparent pixels will be filled to the
color provided in \code{fill} or black (default). If TRUE, image will
maintain the original colors.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
As in \code{\link[ggplot2]{geom_point}}, can be used for scatterplots, but
using raster objects instead of points.
}
\details{
\code{geom_point_pic} differs from \code{\link[ggplot2]{geom_point}} in some
aesthetics. There is no need of \code{colour} aesthetic and the \code{shape}
aesthetic is replaced by the  \code{pic} aesthetic.
}
\section{Aesthetics}{

\code{geom_bar_pic} understand the following aesthetics (required aesthetics
are in bold):
\itemize{
    \item{\bold{x}}
    \item{\bold{y}}
    \item{pic}
    \item{alpha}
    \item{fill}
    \item{size}
    \item{stroke}
}
}
\seealso{
\code{\link[ggplot2]{geom_point}} in the \code{ggplot2} package
}

