% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_pic.R
\name{pic_load_and_fill}
\alias{pic_load_and_fill}
\title{Raster load and fill helper function}
\usage{
pic_load_and_fill(pic, fill, alpha, asis)
}
\arguments{
\item{pic}{character with the raster or matrix object name}

\item{fill}{fill colour, as character}

\item{alpha}{alpha as numeric, from 0 (full transparency) to 1 (full opaque)}

\item{asis}{logical indicating if the pic must be filled or maintained as it
is.}
}
\description{
Loading the raster image and filling int must be done in the draw_group
function inside de GeomBarPic class and also in the draw_key_pic function.
This helper function do that and as a outside function is easy to maintain.
}
\keyword{internal}

