% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_pic.R
\name{scale_pic_manual}
\alias{scale_pic_manual}
\title{Manual scale for ggpic}
\usage{
scale_pic_manual(..., values)
}
\arguments{
\item{...}{common discrete scale parameters: \code{name}, \code{breaks},
\code{labels}, \code{na.value}, \code{limits} and \code{guide}.  See
\code{\link{discrete_scale}} for more details}

\item{values}{a set of aesthetic values to map data values to.  If this
is a named vector, then the values will be matched based on the names.
If unnamed, values will be matched in order (usually alphabetical) with
the limits of the scale.  Any data values that don't match will be
given \code{na.value}.}
}
\description{
Manual scale for pic aesthetic
}
\details{
When mapping pic aesthetic to any categorical or discrete variable, a value
for pic for each level must be provided applying this scale function.
}

